@@ -11,42 +11,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.4",
-    "@angular/cdk": "^7.3.1",
-    "@angular/common": "^7.2.4",
-    "@angular/compiler": "^7.2.4",
-    "@angular/core": "^7.2.4",
-    "@angular/forms": "^7.2.4",
-    "@angular/http": "^7.2.4",
-    "@angular/material": "^7.3.1",
-    "@angular/platform-browser": "^7.2.4",
-    "@angular/platform-browser-dynamic": "^7.2.4",
-    "@angular/router": "^7.2.4",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "core-js": "^3.1.3",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler-cli": "^7.2.4",
-    "@angular/language-service": "^7.2.4",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.800.1",
+    "@angular/cli": "~8.0.1",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@types/jasmine": "~3.3.13",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~12.0.5",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0",
+    "typescript": "^3.4.5"
   }
 }
